requirements.txt
pymongo==4.8.0
pendulum==2.1.2
apache-airflow-providers-mongo==4.2.0
dnspython==2.6.1